+++ tests/Unit/RoutingRedirectorTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Illuminate\Tests\Routing;
+namespace Tests\Unit;
 
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -129,9 +129,9 @@ class RoutingRedirectorTest extends TestCase
     public function testBackRedirectToHttpReferer()
     {
         $this->headers->shouldReceive('has')->with('referer')->andReturn(true);
-        $this->url->shouldReceive('previous')->andReturn('http://foo.com/bar');
+        $this->url->shouldReceive('previous')->andReturn('http://foo.com/bar/');
         $response = $this->redirect->back();
-        $this->assertSame('http://foo.com/bar', $response->getTargetUrl());
+        $this->assertSame('http://foo.com/bar/', $response->getTargetUrl());
     }
 
     public function testAwayDoesntValidateTheUrl()